Updated context menu and dialog functionality
Modified `DialogView.xaml` to add subtitle and conditional subtitle  and message text blocks.
Added new properties in `DialogView.xaml.cs` to manage the appearance of subtitles and messages, and added a `ChangeTextColor` Method for text formatting.

Created a new `DialogText` enum in `DialogText.cs`.

Redesigned `BuildContextMenu` in `ContextMenu.cs` to use named parameters for the settings dialog title, and renamed the "Close" menu item to "Close Application" for greater clarity.

Improved the `Empty` function in `RecycleBinService.cs` to include a confirmation dialog before emptying the Recycle Bin, with a red subtitle.